Generalize Modal components to improve flexibility
